Create preprocessing.py
we have different types of preprocessing. Which works best for a algorithm is hard to say without trying it. So maintain many data views(different transformations) of your data. Then exercise a handful of algorithms on each view of your dataset.
These are reusable codes used to transform the data.